Fix Revit add-in: select/zoom, parameters, Sync Rooms rename, ElementId long
Bug fixes:
- OnexusViewerWindow: fix zoom-to-node silently dropped (only select-node was
  handled); add ShouldZoom to PendingSelect; conditionalize ShowElements call
  so single-click no longer moves the Revit camera; add UniqueId fallback
- Parameters: new BuildElementParameterValueGraph() enumerates element.Parameters
  directly so all built-in, shared, and project parameters are included;
  OnexusParameters now calls this when elements are selected (falling back to
  schema view with a hint when nothing is selected)
- ElementId stored as long throughout (List<int>→List<long> in Schema + all
  export cast sites); OnexusPaneContent/ViewerWindow use long[] + ElementId(long)
- Rename "Sync Spaces" ribbon button to "Sync Rooms" (OST_Rooms are architectural
  rooms, not MEP spaces); updated tooltip

New features:
- BuildRoomsAndElementsSpatialGraph: emit AdjacentTo edges between rooms that
  share a door or window opening (undirected, Spatial dimension)
- Include MEP Spaces (OST_MEPSpaces) alongside architectural rooms in the
  Sync Rooms graph (safe-fail: no-op on architectural-only models)
- Room nodes now carry area (Revit internal sq-ft) and mark parameter values

+++ Revit_Addin/Onexus/OnexusPaneContent.xaml.cs
@@ -178,8 +178,8 @@ namespace Onexus
                         {
                             NodeId     = (string)root["id"],
                             ElementIds = idsJ  != null
-                                ? idsJ.Select(x => (int)x).ToArray()
-                                : Array.Empty<int>(),
+                                ? idsJ.Select(x => (long)x).ToArray()
+                                : Array.Empty<long>(),
                             UniqueIds  = uidsJ != null
                                 ? uidsJ.Select(x => (string)x)
                                        .Where(u => !string.IsNullOrEmpty(u))
@@ -353,7 +353,7 @@ namespace Onexus
                 {
                     var toSelect = new List<ElementId>();
 
-                    // 1) Prefer integer ElementIds — fastest lookup
+                    // 1) Prefer ElementIds — fastest lookup (long for Revit 2024+)
                     if (pending.ElementIds != null && pending.ElementIds.Length > 0)
                     {
                         foreach (var i in pending.ElementIds)
@@ -531,7 +531,7 @@ namespace Onexus
         private class PendingSelect
         {
             public string   NodeId     { get; set; }
-            public int[]    ElementIds { get; set; }
+            public long[]   ElementIds { get; set; }
             public string[] UniqueIds  { get; set; }
             /// <summary>
             /// True for zoom-to-node (right-click menu) — calls ShowElements.
